refactor(control): move udp driver cleanup
